+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Ver. 0.8.3 - 2014-Jun-18
 * Set trollius logger to warning by default (same as asyncio)
 * Added :meth:`.WsgiRequest.redirect` and :meth:`.Router.redirect` methods
   to simplify redirection.
+* Renamed css container as :class:`.Links`
 
 Ver. 0.8.2 - 2014-May-30
 ===========================